refactor model package name
